Introduce a rule-driven system for Series System relationships
The series system defines a number of different relationships that can
exist between records of various types.  Each of these relationships
have some properties in common: a relationship type, start and end
dates, notes, and a relator that specifies the nature of the
relationship.

The RelationshipRules contains the definitions for how different
record types can relate, and sets everything up:

  * For each pair of records participating in a relationship, it
    defines a JSONModel capturing that relationship and provides it to
    the ArchivesSpace DirectionalRelationship definition.

  * It adds properties to the record JSONModel schemas for the new
    relationships.

  * It adds validations to ensure that incoming relationships satisfy
    the rules of the series system.

  * It generates the appropriate sections on the frontend edit &
    readonly forms to allow these new relationship types to be added.

  * It adds new rules to the indexer to make the new relationships
    searchable.