WIP Complicated control grade period calculations
The idea: a record has a list of previous controlling agencies.  An
agency appears in this list if it had exclusive control over the
record in question at any point in time.

ASCII art might help?

Here's a resource with a child AO and a grandchild AO:

      Resource    XXXXXXXXXXXXXXXXXXXXXXXXXXXX
        AO1       YYYYYYYYYYYYYYYZZZZZZZZZZZZZ
          AO2     ----------------------------

X, Y and Z are agencies, and each column represents a day of the
month.  An X in a column means "Agency X controlled this record for
this day", while a dash means control was inherited.

In this example, the top-level resource was controlled by Agency X for
the entire period, but was completely overriden at the child level:
for the first half of the month Agency Y had control; for the second
half Agency Z had control.  As a result, Agency X never had control of
AO1 or AO2.

Here's a small modification:

      Resource    XXXXXXXXXXXXXXXXXXXXXXXXXXXX
        AO1       YYYYYYYYYYYYYY--ZZZZZZZZZZZZ
          AO2     ----------------------------

Now there are two days in the middle of the month where control fell
to Agency X.  They would now have had access for those two days.

Likewise, we could have:

     Resource     XXXXXXXXXXXXXXXXXXXXXXXXXXXX
        AO1       YYYYYYYYYYYYYYZZZZZZZZZZZZZZ
          AO2     --------------X-------------

Giving Agency X control of AO2 for a single day.

If we can establish that a given agency had control of a record at any
point, we record that date and give them 90 days' grace from then on.